init update level01 final
I think that level 1 is now final, so I added a few more spikes in the level after the checkpoint up to the end of the level. I added a new enemy. I added several enemies spread throughout the level. I added other platforms, so for 2 or 3 reasons I have now added color in a different way When you start the game you get 5 lives and even if you die 5 times you get 5 lives globally so now I would say the level 1 who is finished but I also like to accept feedback and add it to level 1 so the who is the final commit